Code generation for rust types (#181)
Closes https://github.com/microsoft/lsprotocol/issues/160

@ebkalderon with this PR most of the types are generated from the spec.
There are some special cases that need to be handled in the generated
code. But, it is good to get a feel for the generated code. I am happy
to review any PR with corrections or suggestions.

* Ignore the `OR` types in this PR. They are temporary.
* Generated type names for "anonymous" types needs work, ignore names
for those in this iteration.
* `extend` (inherited) and `mixins` (basically `Type1 & Type2` in
TypeScript) are implemented by extracting their properties for each
`extend` and `mixin` type and merging them into a single `struct`.
* Quality of life functions (e.g., message directions table)
* Restrictions imposed by LSP on numbers (e.g., unsigned integer must be
in range `[0, 2147483647]` inclusive)
* Most `Option<T>` cases don't need special handling, added skip_if none
for these cases.

Pending (will be covered in followup PRs):
1. Code generation for Request, Notifications, and Responses
1. Special optional in LSP, where absence of a field, filed with `null`
value, and field with non-null value have different meanings. Hence a
simple `Option<T>` is not enough to correctly represent them.